requirement.txt
socksio
aiohttp
tiktoken
openai
google-generativeai
anthropic[vertex]
qianfan
sensenova
reka-api
mistralai
together
cohere
pillow  # for Google image
# spark-ai-python  # for VL
